+++ CHANGELOG.md
@@ -629,6 +629,10 @@
   nesting than necessary.
   ([Giacomo Cavalieri](https://github.com/giacomocavalieri))
 
+- Fixed a bug where the language server wouldn't offer the "unqualify" code
+  action if used on a type alias.
+  ([Giacomo Cavalieri](https://github.com/giacomocavalieri))
+
 - Fixed a bug where the compiler allowed to write a guard with an empty clause.
   ([Tristan-Mihai Radulescu](https://github.com/Courtcircuits))
 
@@ -647,3 +651,7 @@
 - Fixed a bug where the "Extract variable" code action would shadow existing
   variables, constants and function names.
   ([Matias Carlander](https://github.com/matiascr))
+
+- Fixed a bug where the language server would not fill in the missing labels of
+  a pattern correctly, generating invalid code.
+  ([Giacomo Cavalieri](https://github.com/giacomocavalieri))